requirements.txt
asgiref==3.3.1
attrs==20.3.0
Django==3.1.8
django-cors-headers==3.7.0
django-countries==7.1
djangorestframework==3.12.4
django-rest-multiple-models==2.1.3
django-rest-passwordreset==1.1.0
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
requests==2.25.1

